requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.6.0
billiard==3.6.4.0
celery==5.1.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dateutils==0.6.12
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.45.0
django-allauth-2fa==0.8
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-environ==0.8.1
django-filter==21.1
django-lifecycle==0.9.6
django-otp==1.1.1
django-simple-history==3.0.0
django-timezone-field==4.2.1
django-widget-tweaks==1.4.12
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.1
Markdown==3.3.4
msgpack==1.0.3
nsetools==1.0.11
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
pix-apidata==1.3.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyotp==2.6.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
qrcode==7.3.1
redis==4.0.0
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
signalrcore-async==0.5.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
urllib3==1.26.7
urlman==2.0.1
vine==5.0.0
wcwidth==0.2.5
websockets==10.1
wrapt==1.13.3
zope.interface==5.4.0
